fix(cli): move theme engine files to cli
