+++ packages/cli/themes/ThemeProvider.tsx
@@ -1,6 +1,6 @@
 import React, { createContext, useContext, useReducer, useEffect, useCallback } from 'react';
 import type { ThemeConfig, ThemeMode, ThemeState } from '../types/theme';
-import { themeReducer, initialThemeState } from './themeReducer';
+import { themeReducer, initialThemeState } from './ThemeReducer';
 import { injectCSSVariables, removeCSSVariables } from './cssInjector';
 
 interface ThemeContextValue {
