+++ packages/cli/types/theme.ts
@@ -87,3 +87,18 @@ export type ThemeAction =
   | { type: 'RESET_THEME' }
   | { type: 'SET_LOADING'; payload: boolean }
   | { type: 'SET_ERROR'; payload: string | null };
+
+export type ContrastLevel = 'AA' | 'AAA';
+
+export type CreateThemeInput = {
+  id: string;
+  name: string;
+  category: string;
+  primary: string;
+  secondary?: string;
+  accent?: string;
+  description?: string;
+  generateDark?: boolean;
+  contrastLevel?: ContrastLevel;
+  mode?: ThemeMode; // initial generation mode, defaults to light
+};
